add subtotal to bottom of order cart
